Update ContextChecker and incorporate into plan compilation
Update the ContextChecker to check the file paths of FileDataSource and
FileDataSink. Also add checks for Plans with no children but that code
is commented out for now since I'm not sure whether such a situation can
ever occur because a PACT Plan is generated starting from the data
sinks.

Also update MissingChildException and PlanException with a constructor
that allows to specifying a proper error message.

The checkPlan() method is invoked from Client.getOptimizedPlan(), this
way it will always be called when a program is run and the web-client
preview will still work because it circumvents getOptimizedPlan().