Add key-field-less ctor to Builder pattern
Reduce, CoGroup, and Match Builders can now be created without
specifying any key fields. For CoGroup and Match the build() method now
throws an IllegalStateException if now key fields have been set.

Also, rename additionalKeyField() to keyField() because the key field is
not necessarily additional anymore (it might be the only key field.

The changes are mostly due to sopremo, which uses the key less ctor and
then later adds all the keys.