Add getSolutionSet() and .getWorkset() for delta iterations
Before, the result of dataset.iterateDelta() would be the workset and
the solution set could be obtained via iteration.getSolutionSet().

This is confusing since you are not told what the result of
.iterateDelta() is supposed to be. Now, when using the result an
exception is thrown during pre-flight that advises the user to use
the two methods to get the correct dataset (work, solution).