Disable POJO types
There are various issues when working with POJOs in the Java API, for
example [1].

This commit squashes the following commmits:

1. Disable tests, which depend on POJO types

2. Replace expression keys with key selector in test

3. PackagedProgramEndToEndITCase used a KMeans variant, which relies on
expression keys to select the field to group on. This commit replaces
this with a key selector.

4. Disable WordCountPOJO test case and package

5. Disables WordCountPOJOITCase, which executes the WordCountPOJO example.
The example code is still included with the source code (with notice
that it is currently not working), but will *not* be packaged as a JAR
anymore.

6. Disable tests, which depend on POJO types (continued)

[1] https://mail-archives.apache.org/mod_mbox/incubator-flink-dev/201407.mbox/%3C53D96049.1060509%40cse.uta.edu%3E

This closes #93.