[FLINK-1453] Fix failing flink-yarn-tests on OS X
The actual issue was that the web server of the job manager was using the default HTTP port (8081) instead of automatic port assignment by the operating system.
It seems that there are other services running on 8081 on OS X. So the fix is not only for OS X, it applies to all platforms.