[FLINK-1577] Remove misleading error log when failing task externally
Task thread Exceptions were always logged, even if the task was cancelled or
failed externally. In certain scenarios this lead to misleading error messages
and made it hard to figure out the root exception.

When an exception is thrown w/o the task being canceled or failed, the error
is still handled via Task#markFailed(Throwable).