[FLINK-2427] [streaming] Make the BarrierBuffer more robust against lost/missing checkpoint barriers.
Checkpoint barriers are now tolerated to be lost (as may happen if the checkpoint triggering actor
messages are lost). This is realized by allowing the BarrierBuffer to maintain multiple queues of blocked inputs.

The patch also reworks the buffer spilling logic, to increase I/O efficiency, and reduce the main memory
footprint in cases where the buffers have little contents (low flush timeouts).