[FLINK-2769] [runtime-web] Add configurable job manager address to HTTP requests
This was removed in 3b8b4f0f8c0600dc851d676ce1bd7f5ab81cb64f. The
issue was fixed by this, but it disabled local testing as well.

This change re-introduces the variable and sets it to the empty
string by default. This way, we can still use the proxy server for
local testing.

This closes #1449